Move to bukkit package
